# __docs__/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agonez Atlas API
  description: REST API for the Agonez exercise and muscle Atlas and the relational
    PlanCreator draft editor and the desktop Execution workflow API. Authentication
    and ownership are intentionally deferred.
  version: 0.1.0
paths:
  /assets/anatomy.svg:
    get:
      tags:
      - Atlas
      summary: Anatomy Asset
      operationId: anatomy_asset_assets_anatomy_svg_get
      responses:
        '200':
          description: Successful Response
        '404':
          description: Anatomy asset has not been installed
...